fix(dev): BackgroundImage added for the Screens
